Fix fatal bug with basic favoriting system that stops page from loading
It seems that trying to stop the whole site from loading until we look up the id
of every single launcj and then query each individual launch, one at a time...
wasn't exactly the best plan.

Instead of us storing the ids of the favorites and then querying LaunchLibrary
for every id...for every single time we re-render the page,
it would be better to instead simply store the data of the launch already within
the array, and then to simply reload that data whenever the user wants to look
back at their favorites.